pyproject.toml
[project]
name = "proyecto-northwind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "jupyter>=1.1.1",
    "mysql-connector-python>=9.6.0",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.47",
]
